requirements.txt
requests
vcrpy
pytest
python-dotenv